Use destroy signal, not delete-event (last occurences)
It's confusing for users what the delete-event is. It's of use only if
you want to perform some actions like displaying a dialog to save
unsaved documents, or confirmation dialog. Semantically, reacting to the
real destruction of the window, and hence of the main application window
should be handled in the destroy signal, and its name is easier to
remember.